refactor: 將 mock 資料移至 utils 資料夾
